Move reversing/unreversing hash functions from SIMD-code to common code
Non-SIMD builds were failing after 6ed33a7f.

Closes #5264
