+++ src/simd-intrinsics.h
@@ -1,8 +1,11 @@
 /*
- * This software is Copyright (c) 2010 bartavelle, <bartavelle at bandecon.com>, and it is hereby released to the general public under the following terms:
- * Redistribution and use in source and binary forms, with or without modification, are permitted.
- *
- * Some modifications, Jim Fougeron, 2013.  Licensing rights listed in accompanying simd-intrinsics.c file.
+ * This software is
+ * Copyright (c) 2010 bartavelle, <bartavelle at bandecon.com>,
+ * Copyright (c) 2013 JimF,
+ * Copyright (c) 2013-2023 magnum,
+ * and it is hereby released to the general public under the following terms:
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted.
  */
 
 #if !defined (__JTR_SSE_INTRINSICS_H__)
@@ -78,8 +81,6 @@
 #ifdef SIMD_PARA_MD5
 void md5cryptsse(unsigned char *buf, unsigned char *salt, char *out, unsigned int md5_type);
 void SIMDmd5body(vtype* data, uint32_t *out, uint32_t *reload_state, unsigned SSEi_flags);
-void md5_reverse(uint32_t *hash);
-void md5_unreverse(uint32_t *hash);
 #define MD5_ALGORITHM_NAME		BITS " " SIMD_TYPE " " MD5_N_STR
 #else
 #define MD5_ALGORITHM_NAME		"32/" ARCH_BITS_STR
@@ -88,8 +89,6 @@ void md5_unreverse(uint32_t *hash);
 #ifdef SIMD_PARA_MD4
 //void SIMDmd4body(__m128i* data, unsigned int *out, int init);
 void SIMDmd4body(vtype* data, uint32_t *out, uint32_t *reload_state, unsigned SSEi_flags);
-void md4_reverse(uint32_t *hash);
-void md4_unreverse(uint32_t *hash);
 #define MD4_ALGORITHM_NAME		BITS " " SIMD_TYPE " " MD4_N_STR
 #else
 #define MD4_ALGORITHM_NAME		"32/" ARCH_BITS_STR
@@ -97,10 +96,6 @@ void md4_unreverse(uint32_t *hash);
 
 #ifdef SIMD_PARA_SHA1
 void SIMDSHA1body(vtype* data, uint32_t *out, uint32_t *reload_state, unsigned SSEi_flags);
-void sha1_reverse(uint32_t *hash);
-void sha1_unreverse(uint32_t *hash);
-void sha1_reverse3(uint32_t *hash);
-void sha1_unreverse3(uint32_t *hash);
 #define SHA1_ALGORITHM_NAME		BITS " " SIMD_TYPE " " SHA1_N_STR
 #else
 #define SHA1_ALGORITHM_NAME		"32/" ARCH_BITS_STR
@@ -112,18 +107,11 @@ void sha1_unreverse3(uint32_t *hash);
 #ifdef SIMD_COEF_32
 #define SHA256_ALGORITHM_NAME	BITS " " SIMD_TYPE " " SHA256_N_STR
 void SIMDSHA256body(vtype* data, uint32_t *out, uint32_t *reload_state, unsigned SSEi_flags);
-void sha224_reverse(uint32_t *hash);
-void sha224_unreverse(uint32_t *hash);
-void sha256_reverse(uint32_t *hash);
-void sha256_unreverse(uint32_t *hash);
 #endif
 
 #ifdef SIMD_COEF_64
 #define SHA512_ALGORITHM_NAME	BITS " " SIMD_TYPE " " SHA512_N_STR
 void SIMDSHA512body(vtype* data, uint64_t *out, uint64_t *reload_state, unsigned SSEi_flags);
-void sha384_reverse(uint64_t *hash);
-void sha384_unreverse(uint64_t *hash);
-// sha512_reverse is defined in sha2.h. It can be used for a non-SIMD build.
 #endif
 
 #else
